Revert "Detect deployment failures with gateway (#6126)"
This reverts commit 6a9173f962ac9d1fce1c7102ae64b072a517310b.
