+++ test/functional/shared/resources/gateway_test.go
@@ -31,7 +31,6 @@ import (
 	"github.com/radius-project/radius/test/step"
 	"github.com/radius-project/radius/test/validation"
 	"github.com/stretchr/testify/require"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 )
 
 const (
@@ -261,50 +260,6 @@ func Test_Gateway_TLSTermination(t *testing.T) {
 			},
 		},
 	})
-
-	test.Test(t)
-}
-
-func Test_Gateway_Failure(t *testing.T) {
-	template := "testdata/corerp-resources-gateway-failure.bicep"
-	name := "corerp-resources-gateway-failure"
-	secret := "secret"
-
-	// We might see either of these states depending on the timing.
-	validateFn := step.ValidateAnyDetails("DeploymentFailed", []step.DeploymentErrorDetail{
-		{
-			Code: "ResourceDeploymentFailure",
-			Details: []step.DeploymentErrorDetail{
-				{
-					Code:            "Internal",
-					MessageContains: "invalid TLS certificate",
-				},
-			},
-		},
-	})
-
-	test := shared.NewRPTest(t, name, []shared.TestStep{
-		{
-			Executor:                               step.NewDeployErrorExecutor(template, validateFn),
-			SkipObjectValidation:                   true,
-			SkipKubernetesOutputResourceValidation: true,
-		},
-	},
-		unstructured.Unstructured{
-			Object: map[string]interface{}{
-				"apiVersion": "v1",
-				"kind":       "Secret",
-				"metadata": map[string]interface{}{
-					"name":      secret,
-					"namespace": "mynamespace",
-				},
-				"type": "Opaque",
-				"data": map[string]interface{}{
-					"tls.crt": "",
-					"tls.key": "",
-				},
-			},
-		})
 
 	test.Test(t)
 }
